+++ models/contacts.json
@@ -60,9 +60,9 @@
     "phone": "(748) 206-2688"
   },
   {
-    "name": "Jon2",
+    "name": "don",
     "email": "aaaaaaae@fdsfdsf.com",
-    "phone": "0660336611",
+    "phone": "1111111111",
     "id": "1692552573483"
   }
 ]
